fix: Fixed synchronize timeout in test memory multithread
Fix for the test failures observed in Simulation environment
with long execution times due to zeCommandSynchronize
returning ZE_RESULT_NOT_READY.
zeCommandQueueSynchronize shall be called with UINT64_MAX
for the function to not return until complete or device is lost
(according to oneAPI Level Zero spec).

Related-To: NEO-10446

+++ conformance_tests/core/test_multithread/src/test_memory_multithread.cpp
@@ -1,6 +1,6 @@
 /*
  *
- * Copyright (C) 2020 Intel Corporation
+ * Copyright (C) 2024 Intel Corporation
  *
  * SPDX-License-Identifier: MIT
  *
@@ -236,7 +236,7 @@ void perform_memory_manipulation() {
     lzt::append_memory_copy(cmd_list, host_mem, dev_mem, xfer_size, nullptr);
     lzt::close_command_list(cmd_list);
     lzt::execute_command_lists(cq, 1, &cmd_list, nullptr);
-    lzt::synchronize(cq, UINT32_MAX);
+    lzt::synchronize(cq, UINT64_MAX);
     lzt::event_host_synchronize(event_barrier_to_host, UINT32_MAX - 1);
     lzt::query_event(event_barrier_to_host);
